requirements.txt
# SecureRAG v1.0 Dependencies

# MCP Server Framework
fastmcp>=0.2.0

# Vector Database
qdrant-client>=1.7.0

# ML/Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# Document Processing
pymupdf>=1.23.0
python-docx>=0.8.11
pypdf>=3.0.0

# Text Processing
langchain-text-splitters>=0.0.1

# Security
cryptography>=41.0.0

# Configuration & Data
pyyaml>=6.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# Optional Cloud Providers
openai>=1.0.0
anthropic>=0.18.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0
